Refactor source to avoid circular imports, which Python3.x does not like. This patch moves Interface and BaseInterface to aciphysobjects. Both classes are imported in acitoolkit.py so any existing code doing 'from acitoolkit.acitoolkit import Interface' will still work.
