+++ acitoolkit/aciphysobject.py
@@ -18,8 +18,7 @@
 
 from .acibaseobject import BaseACIObject
 from .acisession import Session
-import acitoolkit as ACI
-from .acicounters import AtomicCountersOnGoing
+from .acicounters import AtomicCountersOnGoing, InterfaceStats
 
 import logging
 import re
@@ -537,7 +536,7 @@ class Linecard(BaseACIPhysModule):
         """
 
         # The following will add the interfaces to the linecard
-        ACI.Interface.get(self._session, self)
+        Interface.get(self._session, self)
         
         if deep:
             for child in self._children:
@@ -1792,7 +1791,7 @@ class Link(BaseACIPhysObject):
             raise TypeError("Parent pod must be specified in order to get node")
         linecard = self.get_slot1()
         if linecard:
-            interfaces = linecard.get_children(ACI.Interface)
+            interfaces = linecard.get_children(Interface)
             for interface in interfaces:
                 if interface.port == self.port1:
                     return interface
@@ -1813,7 +1812,7 @@ class Link(BaseACIPhysObject):
                              "order to get node"))
         linecard = self.get_slot2()
         if linecard:
-            interfaces = linecard.get_children(ACI.Interface)
+            interfaces = linecard.get_children(Interface)
             for interface in interfaces:
                 if interface.port == self.port2:
                     return interface
@@ -1863,5 +1862,546 @@ class AccessPolicyConcrete(BaseACIObject):
         self.direction = None
         self.filter_id = None
         self.mask_dscp = None
-        
-    
+
+class BaseInterface(BaseACIObject):
+    """Abstract class used to provide base functionality to other Interface
+       classes.
+    """
+    def _get_port_selector_json(self, port_type, port_name):
+        """Returns the json used for selecting the specified interfaces
+        """
+        name = self._get_name_for_json()
+        port_blk = {'name': name,
+                    'fromCard': self.module,
+                    'toCard': self.module,
+                    'fromPort': self.port,
+                    'toPort': self.port}
+        port_blk = {'infraPortBlk': {'attributes': port_blk,
+                                     'children': []}}
+        pc_url = 'uni/infra/funcprof/%s-%s' % (port_type, port_name)
+        accbasegrp = {'infraRsAccBaseGrp': {'attributes': {'tDn': pc_url},
+                                            'children': []}}
+        portselect = {'infraHPortS': {'attributes': {'name': name,
+                                                     'type': 'range'},
+                                      'children': [port_blk, accbasegrp]}}
+        accport_selector = {'infraAccPortP': {'attributes': {'name': name},
+                                              'children': [portselect]}}
+        node_blk = {'name': name,
+                    'from_': self.node, 'to_': self.node}
+        node_blk = {'infraNodeBlk': {'attributes': node_blk, 'children': []}}
+        leaf_selector = {'infraLeafS': {'attributes': {'name': name,
+                                                       'type': 'range'},
+                                        'children': [node_blk]}}
+        accport = {'infraRsAccPortP':
+                   {'attributes': {'tDn': 'uni/infra/accportprof-%s' % name},
+                    'children': []}}
+        node_profile = {'infraNodeP': {'attributes': {'name': name},
+                                       'children': [leaf_selector,
+                                                    accport]}}
+        return node_profile, accport_selector
+
+    def get_port_selector_json(self):
+        return self._get_port_selector_json('accportgrp',
+                                            self._get_name_for_json())
+
+    def get_port_channel_selector_json(self, port_name):
+        return self._get_port_selector_json('accbundle', port_name)
+
+class Interface(BaseInterface):
+    """This class defines a physical interface.
+    """
+    def __init__(self, interface_type, pod, node, module, port,
+                 parent=None, session=None, attributes={}):
+
+        self._session = session
+        self.attributes = {}
+        self.attributes = copy.deepcopy(attributes)
+        self.interface_type = str(interface_type)
+        self.pod = str(pod)
+        self.node = str(node)
+        self.module = str(module)
+        self.port = str(port)
+        self.attributes['interface_type'] = str(interface_type)
+        self.attributes['pod'] = str(pod)
+        self.attributes['node'] = str(node)
+        self.attributes['module'] = str(module)
+        self.attributes['port'] = str(port)
+
+        self.if_name = self.interface_type + ' ' + self.pod + '/'
+        self.if_name += self.node + '/' + self.module + '/' + self.port
+        self.attributes['if_name'] = self.if_name
+        super(Interface, self).__init__(self.if_name, None)
+        self.porttype = ''
+        self.adminstatus = ''    # up or down
+        self.speed = '10G'       # 100M, 1G, 10G or 40G
+        self.mtu = ''
+        self._cdp_config = None
+        self._lldp_config = None
+        self.type = 'interface'
+        self.attributes['type'] = 'interface'
+        self.id = interface_type + module + '/' + port
+
+        self._parent = parent
+        if parent:
+            self._parent.add_child(self)
+        self.stats = InterfaceStats(self, self.attributes.get('dist_name'))
+
+    def is_interface(self):
+        """
+        Returns whether this instance is considered an interface.
+
+        :returns: True
+        """
+        return True
+
+    def is_cdp_enabled(self):
+        """
+        Returns whether this interface has CDP configured as enabled.
+
+        :returns: True or False
+        """
+        return self._cdp_config == 'enabled'
+
+    def is_cdp_disabled(self):
+        """
+        Returns whether this interface has CDP configured as disabled.
+
+        :returns: True or False
+        """
+        return self._cdp_config == 'disabled'
+
+    def enable_cdp(self):
+        """
+        Enables CDP on this interface.
+        """
+        self._cdp_config = 'enabled'
+
+    def disable_cdp(self):
+        """
+        Disables CDP on this interface.
+        """
+        self._cdp_config = 'disabled'
+
+    def is_lldp_enabled(self):
+        """
+        Returns whether this interface has LLDP configured as enabled.
+
+        :returns: True or False
+        """
+        return self._lldp_config == 'enabled'
+
+    def is_lldp_disabled(self):
+        """
+        Returns whether this interface has LLDP configured as disabled.
+
+        :returns: True or False
+        """
+        return self._lldp_config == 'disabled'
+
+    def enable_lldp(self):
+        """
+        Enables LLDP on this interface.
+        """
+        self._lldp_config = 'enabled'
+
+    def disable_lldp(self):
+        """
+        Disables LLDP on this interface.
+        """
+        self._lldp_config = 'disabled'
+
+    def get_type(self):
+        return self.type
+
+    def get_serial(self):
+        return None
+
+    def get_url(self):
+        phys_domain_url = '/api/mo/uni.json'
+        fabric_url = '/api/mo/uni/fabric.json'
+        infra_url = '/api/mo/uni.json'
+        return phys_domain_url, fabric_url, infra_url
+
+    def _get_name_for_json(self):
+        return '%s-%s-%s-%s' % (self.pod, self.node,
+                                self.module, self.port)
+
+    def get_json(self):
+        """ Get the json for an interface.  Returns a tuple since the json is
+            required to be sent in 2 posts.
+        """
+        fabric = None
+        # Physical Domain json
+        vlan_ns_dn = 'uni/infra/vlanns-allvlans-static'
+        vlan_ns_ref = {'infraRsVlanNs': {'attributes':
+                                         {'tDn': vlan_ns_dn},
+                                         'children': []}}
+        phys_domain = {'physDomP': {'attributes': {'name': 'allvlans'},
+                                    'children': [vlan_ns_ref]}}
+
+        # Infra json
+        infra = {'infraInfra': {'children': []}}
+        node_profile, accport_selector = self.get_port_selector_json()
+        infra['infraInfra']['children'].append(node_profile)
+        infra['infraInfra']['children'].append(accport_selector)
+        speed_name = 'speed%s' % self.speed
+        hifpol_dn = 'uni/infra/hintfpol-%s' % speed_name
+        speed = {'fabricHIfPol': {'attributes': {'autoNeg': 'on',
+                                                 'dn': hifpol_dn,
+                                                 'name': speed_name,
+                                                 'speed': self.speed},
+                                  'children': []}}
+        infra['infraInfra']['children'].append(speed)
+        name = self._get_name_for_json()
+        accportgrp_dn = 'uni/infra/funcprof/accportgrp-%s' % name
+        speed_attr = {'tnFabricHIfPolName': speed_name}
+        speed_children = {'infraRsHIfPol': {'attributes': speed_attr,
+                                            'children': []}}
+        cdp_children = None
+        if self._cdp_config is not None:
+            cdp_data = {'tnCdpIfPolName': 'CDP_%s' % self._cdp_config}
+            cdp_children = {'infraRsCdpIfPol': {'attributes': cdp_data}}
+        lldp_children = None
+        if self._lldp_config is not None:
+            lldp_data = {'tnLldpIfPolName': 'LLDP_%s' % self._lldp_config}
+            lldp_children = {'infraRsLldpIfPol': {'attributes': lldp_data}}
+        att_ent_dn = 'uni/infra/attentp-allvlans'
+        att_ent_p = {'infraRsAttEntP': {'attributes': {'tDn': att_ent_dn},
+                                        'children': []}}
+        speed_ref = {'infraAccPortGrp': {'attributes': {'dn': accportgrp_dn,
+                                                        'name': name},
+                                         'children': [speed_children,
+                                                      att_ent_p]}}
+        if cdp_children is not None:
+            speed_ref['infraAccPortGrp']['children'].append(cdp_children)
+        if lldp_children is not None:
+            speed_ref['infraAccPortGrp']['children'].append(lldp_children)
+        speed_ref = {'infraFuncP': {'attributes': {}, 'children': [speed_ref]}}
+        infra['infraInfra']['children'].append(speed_ref)
+
+        phys_dom_dn = 'uni/phys-allvlans'
+        rs_dom_p = {'infraRsDomP': {'attributes': {'tDn': phys_dom_dn}}}
+        infra_att_entity_p = {'infraAttEntityP': {'attributes':
+                                                  {'name': 'allvlans'},
+                                                  'children': [rs_dom_p]}}
+        infra['infraInfra']['children'].append(infra_att_entity_p)
+
+        if self._cdp_config is not None:
+            cdp_if_pol = {'cdpIfPol': {'attributes': {'adminSt': self._cdp_config,
+                                                      'name': 'CDP_%s' % self._cdp_config}}}
+            infra['infraInfra']['children'].append(cdp_if_pol)
+
+        if self._lldp_config is not None:
+            lldp_if_pol = {'lldpIfPol': {'attributes': {'adminRxSt': self._lldp_config,
+                                                        'adminTxSt': self._lldp_config,
+                                                        'name': 'LLDP_%s' % self._lldp_config}}}
+            infra['infraInfra']['children'].append(lldp_if_pol)
+
+        if self.adminstatus != '':
+            adminstatus_attributes = {}
+            adminstatus_attributes['tDn'] = self._get_path()
+            if self.adminstatus == 'up':
+                admin_dn = 'uni/fabric/outofsvc/rsoosPath-['
+                admin_dn = admin_dn + self._get_path() + ']'
+                adminstatus_attributes['dn'] = admin_dn
+                adminstatus_attributes['status'] = 'deleted'
+            else:
+                adminstatus_attributes['lc'] = 'blacklist'
+            adminstatus_json = {'fabricRsOosPath':
+                                {'attributes': adminstatus_attributes,
+                                 'children': []}}
+            fabric = {'fabricOOServicePol': {'children': [adminstatus_json]}}
+
+        fvns_encap_blk = {'fvnsEncapBlk': {'attributes': {'name': 'encap',
+                                                          'from': 'vlan-1',
+                                                          'to': 'vlan-4092'}}}
+        fvns_vlan_inst_p = {'fvnsVlanInstP': {'attributes':
+                                              {'name': 'allvlans',
+                                               'allocMode': 'static'},
+                                              'children': [fvns_encap_blk]}}
+        infra['infraInfra']['children'].append(fvns_vlan_inst_p)
+
+        return phys_domain, fabric, infra
+
+    def _get_path(self):
+        """Get the path of this interface used when communicating with
+           the APIC object model.
+        """
+        return 'topology/pod-%s/paths-%s/pathep-[eth%s/%s]' % (self.pod,
+                                                               self.node,
+                                                               self.module,
+                                                               self.port)
+
+    @staticmethod
+    def parse_name(name):
+        """Parses a name that is of the form:
+        <type> <pod>/<mod>/<port>
+        """
+        interface_type = name.split()[0]
+        name = name.split()[1]
+        (pod, node, module, port) = name.split('/')
+        return interface_type, pod, node, module, port
+
+    @staticmethod
+    def _parse_physical_dn(dn):
+        """
+        Handles DNs that look like the following:
+        topology/pod-1/node-103/sys/phys-[eth1/12]
+        """
+        name = dn.split('/')
+        pod = name[1].split('-')[1]
+        node = name[2].split('-')[1]
+        module = name[4].split('[')[1]
+        interface_type = module[:3]
+        module = module[3:]
+        port = name[5].split(']')[0]
+
+        return interface_type, pod, node, module, port
+
+    @staticmethod
+    def _parse_path_dn(dn):
+        """
+        Handles DNs that look like the following:
+        topology/pod-1/paths-102/pathep-[eth1/12]
+        """
+        name = dn.split('/')
+        pod = name[1].split('-')[1]
+        node = name[2].split('-')[1]
+        module = name[3].split('[')[1]
+        interface_type = module[:3]
+        module = module[3:]
+        port = name[4].split(']')[0]
+
+        return interface_type, pod, node, module, port
+
+    @classmethod
+    def parse_dn(cls, dn):
+        """
+        Parses the pod, node, module, port from a distinguished name
+        of the interface.
+
+        :param dn: String containing the interface distinguished name
+        :returns: interface_type, pod, node, module, port
+        """
+        if 'sys' in dn.split('/'):
+            return cls._parse_physical_dn(dn)
+        else:
+            return cls._parse_path_dn(dn)
+
+    @staticmethod
+    def _get_discoveryprot_policies(session, prot):
+        """
+        :param prot: String containing either 'cdp' or 'lldp'
+        """
+        prot_policies = {}
+        if prot == 'cdp':
+            prot_class = 'cdpIfPol'
+        elif prot == 'lldp':
+            prot_class = 'lldpIfPol'
+        else:
+            raise ValueError
+
+        query_url = '/api/node/class/%s.json?query-target=self' % prot_class
+        ret = session.get(query_url)
+        prot_data = ret.json()['imdata']
+        for policy in prot_data:
+            attributes = policy['%s' % prot_class]['attributes']
+            if prot == 'cdp':
+                prot_policies[attributes['name']] = attributes['adminSt']
+            else:
+                prot_policies[attributes['name']] = attributes['adminTxSt']
+        return prot_policies
+
+    @staticmethod
+    def _get_discoveryprot_relations(session, interfaces, prot, prot_policies):
+        if prot == 'cdp':
+            prot_relation_class = 'l1RsCdpIfPolCons'
+            prot_relation_dn_class = '/cdpIfP-'
+            prot_relation_dn = '/rscdpIfPolCons'
+        elif prot == 'lldp':
+            prot_relation_class = 'l1RsLldpIfPolCons'
+            prot_relation_dn_class = '/lldpIfP-'
+            prot_relation_dn = '/rslldpIfPolCons'
+        else:
+            raise ValueError
+
+        query_url = ('/api/node/class/l1PhysIf.json?query-target=subtree&'
+                     'target-subtree-class=%s' % prot_relation_class)
+        ret = session.get(query_url)
+        prot_data = ret.json()['imdata']
+        for prot_relation in prot_data:
+            attributes = prot_relation[prot_relation_class]['attributes']
+            policy_name = attributes['tDn'].split(prot_relation_dn_class)[1]
+            intf_dn = attributes['dn'].split(prot_relation_dn)[0]
+            search_intf = Interface(*Interface._parse_physical_dn(intf_dn))
+            for intf in interfaces:
+                if intf == search_intf:
+                    if prot_policies[policy_name] == 'enabled':
+                        if prot == 'cdp':
+                            intf.enable_cdp()
+                        else:
+                            intf.enable_lldp()
+                    else:
+                        if prot == 'cdp':
+                            intf.disable_cdp()
+                        else:
+                            intf.disable_lldp()
+                    break
+        return interfaces
+
+    @staticmethod
+    def get(session, pod_parent=None, node=None, module=None, port=None):
+        """
+        Gets all of the physical interfaces from the APIC if no parent is
+        specified. If a parent of type Linecard is specified, then only
+        those interfaces on that linecard are returned and they are also
+        added as children to that linecard.
+
+        If the pod, node, module and port are specified, then only that
+        specific interface is read.
+
+        :param session: the instance of Session used for APIC communication
+        :param pod_parent: Linecard instance to limit interfaces or pod\
+                           number (optional)
+        :param node: Node id string.  This specifies the switch to read.\
+                     (optional)
+        :param module: Module id string.  This specifies the module or\
+                       slot of the port. (optional)
+        :param port: Port number.  This is the port to read. (optional)
+
+        :returns: list of Interface instances
+        """
+
+        if port:
+            if not isinstance(port, str) :
+                raise TypeError('When specifying a specific port, the port'
+                                ' must be a identified by a string')
+            if not isinstance(module, str):
+                raise TypeError(('When specifying a specific port, the module'
+                                  ' must be identified by a string'))
+            if not isinstance(node, str):
+                raise TypeError(('When specifying a specific port, the node '
+                                  'must be identified by a string'))
+            if not isinstance(pod_parent, str):
+                raise TypeError(('When specifying a specific port, the pod '
+                                  'must be identified by a string'))
+            parent = None
+        else :
+            if pod_parent :
+                if not isinstance(pod_parent, Linecard) :
+                    raise TypeError('Interface parent must be a linecard object')
+
+            parent = pod_parent
+
+        if not isinstance(session, Session):
+            raise TypeError('An instance of Session class is required')
+
+        cdp_policies = Interface._get_discoveryprot_policies(session, 'cdp')
+        lldp_policies = Interface._get_discoveryprot_policies(session, 'lldp')
+
+        if port:
+            dist_name = 'topology/pod-{0}/node-{1}/sys/phys-[eth{2}/{3}]'.format(pod_parent, node, module, port)
+            interface_query_url = ('/api/mo/' + dist_name + '.json?query-target=self')
+            eth_query_url = ('/api/mo/' + dist_name + '/phys.json?query-target=self')
+        else:
+            interface_query_url = ('/api/node/class/l1PhysIf.json?query-target='
+                                   'self')
+            eth_query_url = ('/api/node/class/ethpmPhysIf.json?query-target='
+                             'self')
+
+        ret = session.get(interface_query_url)
+        resp = []
+        interface_data = ret.json()['imdata']
+
+        # also get information about the ethernet interface
+        ethResp = session.get(eth_query_url)
+        resp = []
+        ethData = ethResp.json()['imdata']
+
+        # re-index the ethernet port info so it can be referenced by dn
+        ethDataDict = {}
+        for object in ethData:
+
+            ethDataDict[object['ethpmPhysIf']['attributes']['dn']] = object['ethpmPhysIf']['attributes']
+
+        for interface in interface_data:
+            attributes = {}
+            dist_name = str(interface['l1PhysIf']['attributes']['dn'])
+            attributes['dist_name'] = dist_name
+            porttype = str(interface['l1PhysIf']['attributes']['portT'])
+            attributes['porttype'] = porttype
+            adminstatus = str(interface['l1PhysIf']['attributes']['adminSt'])
+            attributes['adminstatus'] = adminstatus
+            speed = str(interface['l1PhysIf']['attributes']['speed'])
+            attributes['speed'] = speed
+            mtu = str(interface['l1PhysIf']['attributes']['mtu'])
+            attributes['mtu'] = mtu
+            id = str(interface['l1PhysIf']['attributes']['id'])
+            attributes['id'] = id
+            attributes['monPolDn'] = str(interface['l1PhysIf']['attributes']['monPolDn'])
+            attributes['name'] = str(interface['l1PhysIf']['attributes']['name'])
+            attributes['descr'] = str(interface['l1PhysIf']['attributes']['descr'])
+            attributes['usage'] = str(interface['l1PhysIf']['attributes']['usage'])
+            (interface_type, pod, node,
+             module, port) = Interface.parse_dn(dist_name)
+            attributes['interface_type'] = interface_type
+            attributes['pod'] = pod
+            attributes['node'] = node
+            attributes['module'] = module
+            attributes['port'] = port
+            attributes['operSt'] = ethDataDict[dist_name + '/phys']['operSt']
+            interface_obj = Interface(interface_type, pod, node, module, port, parent=None, session=session, attributes=attributes)
+            interface_obj.porttype = porttype
+            interface_obj.adminstatus = adminstatus
+            interface_obj.speed = speed
+            interface_obj.mtu = mtu
+
+            if not isinstance(pod_parent, str) and pod_parent:
+                if interface_obj.pod == pod_parent.pod and interface_obj.node == pod_parent.node and interface_obj.module == pod_parent.slot:
+                    interface_obj._parent = pod_parent
+                    interface_obj._parent.add_child(interface_obj)
+                    resp.append(interface_obj)
+            else:
+                resp.append(interface_obj)
+
+        resp = Interface._get_discoveryprot_relations(session, resp, 'cdp', cdp_policies)
+        resp = Interface._get_discoveryprot_relations(session, resp, 'lldp', lldp_policies)
+        return resp
+
+    def __str__(self):
+        items = [self.if_name, '\t', self.porttype, '\t',
+                 self.adminstatus, '\t', self.speed, '\t',
+                 self.mtu]
+        ret = ''.join(items)
+        return ret
+
+    def __eq__(self, other):
+        if type(self) != type(other):
+            return False
+        if (self.attributes['interface_type'] == other.attributes.get('interface_type') and
+                self.attributes['pod'] == other.attributes.get('pod') and
+                self.attributes['node'] == other.attributes.get('node') and
+                self.attributes['module'] == other.attributes.get('module') and
+                self.attributes['port'] == other.attributes.get('port')):
+
+            return True
+        return False
+    def get_adjacent_port(self) :
+        """
+        This will return the port ID of the port at the other end of the link.
+
+        For Access ports, it will only have a result if it is connected to
+        a controller node.
+
+        If no link is found, then the result will be None.  That does not mean
+        that nothing is connected, just that a fabric link is not connected.
+
+        :returns : Port ID string
+        """
+        result = None
+
+        links = Link.get(self._session,'1',self.attributes['node'])
+        for link in links :
+            if link.port1==self.attributes['port'] :
+                return link.get_port_id2()
+        return result
